Use seperated pages to avoid binding failure for elements not in logical tree. Re-arrange some utils since standalone file in CommnUtil may cause multiple instance among different libraries, causing conflicts. Some code clean. [skip ci]
