+++ AnyDock/AnyDockHost.res.xaml
@@ -2,20 +2,4 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:wpf="clr-namespace:System.Windows.Controls;assembly=PresentationFramework" 
                     xmlns:local="clr-namespace:AnyDock">
-    <ControlTemplate x:Key="AnyDockHostTemplate" TargetType="{x:Type local:AnyDockHost}">
-        <!--, diag:PresentationTraceSources.TraceLevel=High-->
-        <DockPanel Background="{TemplateBinding Background}">
-            <local:AnyDockSidePanel x:Name="LeftPanel"   DockPanel.Dock="Left"   TabStripPlacement="Left"   
-                                    Background="{TemplateBinding Background}" MaxWidth ="400"/>
-            <local:AnyDockSidePanel x:Name="RightPanel"  DockPanel.Dock="Right"  TabStripPlacement="Right"  
-                                    Background="{TemplateBinding Background}" MaxWidth ="400"/>
-            <local:AnyDockSidePanel x:Name="TopPanel"    DockPanel.Dock="Top"    TabStripPlacement="Top"    
-                                    Background="{TemplateBinding Background}" MaxHeight="200"/>
-            <local:AnyDockSidePanel x:Name="BottomPanel" DockPanel.Dock="Bottom" TabStripPlacement="Bottom" 
-                                    Background="{TemplateBinding Background}" MaxHeight="200"/>
-            <ContentPresenter x:Name="CenterPanel"/>
-        </DockPanel>
-        <ControlTemplate.Triggers>
-        </ControlTemplate.Triggers>
-    </ControlTemplate>
 </ResourceDictionary>
