+++ CommonUtil/CommonUtil.projitems
@@ -11,7 +11,6 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)BindingHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MemoryMonitor.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ProxyBinder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ViewModelHelper.cs" />
   </ItemGroup>
 </Project>
