--- WPFTest/DrawablePage.xaml
@@ -1,18 +1,0 @@
-﻿<UserControl x:Class="WPFTest.DrawablePage"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:WPFTest"
-             xmlns:ext="clr-namespace:XZiar.WPFControl;assembly=WPFControlExt"
-             xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
-             mc:Ignorable="d" 
-             d:DesignHeight="800" d:DesignWidth="400"
-             Background="{Binding Background, RelativeSource={RelativeSource FindAncestor, AncestorType=TabControl}}">
-    <DockPanel>
-        <ComboBox x:Name="cboxDrawable" Background="{StaticResource brshDarkBackground}" Foreground="White" Template="{StaticResource darkCombobox}" Margin="0,0,0,8" DockPanel.Dock="Top" ItemsSource="{Binding DataContext, RelativeSource={RelativeSource FindAncestor, AncestorType=UserControl}, Mode=OneWay}" />
-        <!--<xctk:PropertyGrid x:Name="pgDrawable" AutoGenerateProperties="True" SelectedObject="{Binding SelectedItem, ElementName=cboxDrawable, Mode=OneWay}">
-        </xctk:PropertyGrid>-->
-        <local:ControllableGrid x:Name="ctlDrawable" Background="{StaticResource brshDarkBackground}" Foreground="White" DataContext="{Binding SelectedItem, ElementName=cboxDrawable, Mode=OneWay}" />
-    </DockPanel>
-</UserControl>
