--- WPFTest/LightPage.xaml
@@ -1,68 +1,0 @@
-﻿<UserControl x:Class="WPFTest.LightPage"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:WPFTest"
-             xmlns:ext="clr-namespace:XZiar.WPFControl;assembly=WPFControlExt"
-             xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
-             mc:Ignorable="d" 
-             d:DesignHeight="800" d:DesignWidth="400"
-             Background="{Binding Background, RelativeSource={RelativeSource FindAncestor, AncestorType=TabControl}}">
-    <DockPanel>
-        <ComboBox x:Name="cboxLight" Background="{StaticResource brshDarkBackground}" Foreground="White" Template="{StaticResource darkCombobox}" Margin="0,0,0,8" DockPanel.Dock="Top" ItemsSource="{Binding DataContext, RelativeSource={RelativeSource FindAncestor, AncestorType=UserControl}, Mode=OneWay}"/>
-        <local:ControllableGrid x:Name="ctlLight" Background="{StaticResource brshDarkBackground}" Foreground="White" DataContext="{Binding SelectedItem, ElementName=cboxLight, Mode=OneWay}" />
-        <!--<ScrollViewer Margin="0" CanContentScroll="False" PanningMode="VerticalOnly" VerticalScrollBarVisibility="Hidden">
-            <StackPanel x:Name="stkLight" Orientation="Vertical" DataContext="{Binding SelectedItem, ElementName=cboxLight, Mode=OneWay}">
-                <ext:LabelControl Foreground="White" Background="{StaticResource brshDarkBackground}" LabelWidth="64" Label="启用">
-                    <CheckBox IsChecked="{Binding IsOn, Mode=TwoWay}" />
-                </ext:LabelControl>
-                <ext:LabelTextBox Foreground="White" Background="{StaticResource brshDarkBackground}" Content="{Binding Type, Mode=OneTime}" LabelWidth="64" Label="类型" IsReadOnly="True"/>
-                <ext:LabelTextBox Foreground="White" Background="{StaticResource brshDarkBackground}" Content="{Binding Name, Mode=TwoWay}" LabelWidth="64" Label="名称" IsReadOnly="False"/>
-                <ext:LabelTextBox Foreground="White" Background="{StaticResource brshDarkBackground}" Content="{Binding Position, Mode=OneWay}" LabelWidth="64" Label="位置" IsReadOnly="True">
-                    <ext:LabelTextBox.Style>
-                        <Style TargetType="ext:LabelTextBox">
-                            <Style.Triggers>
-                                <DataTrigger Binding="{Binding Type, Mode=OneWay}" Value="Parallel">
-                                    <Setter Property="Visibility" Value="Collapsed"/>
-                                </DataTrigger>
-                            </Style.Triggers>
-                        </Style>
-                    </ext:LabelTextBox.Style>
-                </ext:LabelTextBox>
-                <ext:LabelTextBox Foreground="White" Background="{StaticResource brshDarkBackground}" Content="{Binding Direction, Mode=OneWay}" LabelWidth="64" Label="方向" IsReadOnly="True">
-                    <ext:LabelTextBox.Style>
-                        <Style TargetType="ext:LabelTextBox">
-                            <Style.Triggers>
-                                <DataTrigger Binding="{Binding Type, Mode=OneTime}" Value="Point">
-                                    <Setter Property="Visibility" Value="Collapsed"/>
-                                </DataTrigger>
-                            </Style.Triggers>
-                        </Style>
-                    </ext:LabelTextBox.Style>
-                </ext:LabelTextBox>
-                <ext:LabelControl Foreground="White" Background="{StaticResource brshDarkBackground}" LabelWidth="64" Label="颜色">
-                    <xctk:ColorPicker SelectedColor="{Binding Color, Mode=TwoWay}" ColorMode="ColorCanvas"/>
-                </ext:LabelControl>
-                <ext:LabelControl Foreground="White" Background="{StaticResource brshDarkBackground}" LabelWidth="80" Label="亮度">
-                    <xctk:DoubleUpDown Foreground="White" Value="{Binding Luminance}" FormatString="F2" Increment=".1"/>
-                </ext:LabelControl>
-                <ext:LabelControl Foreground="White" Background="{StaticResource brshDarkBackground}" LabelWidth="64" Label="切光角">
-                    <ext:LabelControl.Style>
-                        <Style TargetType="ext:LabelControl">
-                            <Style.Triggers>
-                                <DataTrigger Binding="{Binding Type, Mode=OneTime}" Value="Parallel">
-                                    <Setter Property="Visibility" Value="Collapsed"/>
-                                </DataTrigger>
-                                <DataTrigger Binding="{Binding Type, Mode=OneTime}" Value="Point">
-                                    <Setter Property="Visibility" Value="Collapsed"/>
-                                </DataTrigger>
-                            </Style.Triggers>
-                        </Style>
-                    </ext:LabelControl.Style>
-                    <xctk:RangeSlider LowerValue="{Binding CutoffInner, Mode=TwoWay}" HigherValue="{Binding CutoffOuter, Mode=TwoWay}" Minimum="0" Maximum="180" AutoToolTipPlacement="TopLeft" AutoToolTipPrecision="1" VerticalContentAlignment="Center"/>
-                </ext:LabelControl>
-            </StackPanel>
-        </ScrollViewer>-->
-    </DockPanel>
-</UserControl>
